Use generated data instead of "California housing" from scikit-learn.
California housing has to be downloaded from the Internet, and we would
like to remove as many of these dependencies as possible. For these
tests we don't really need real data, so generating it is fine.

Disable the both "fit_modes_all_return_equal_results" tests as the fit
modes can actually return quite different results due to different
management of the random state. The tests were only passing by chance
before, and we have had issues with them before.