Use a Generator object to specify the seed for the random embeddings.
Currently we set the global pytorch random seed. When multi-device
inference is enabled, this can result in a race condition because
the global seed is shared between threads. Use a Generator instead to
avoid this.

Exporting the generator is not supported by onnx, so don't use it during
tracing. This means that the random embeddings will be different to
those used during training, which may effect the performance of the
model. However, the previous method for fixing the seed was silently
ignored by onnx export, so it doesn't make things worse.

Testing:
- Save the sampled embeddings from the old and the new versions and
  check they are equal.
- Existing consistency checks, plus the additional two I have added to
  verify multi-device inference.