Specify numpy as a direct dependency.
Our code directly imports numpy but we don't delcare it as an explicit
dependency, so we should do this.

This has the side-effect of fixing the test failures seen in
https://github.com/PriorLabs/TabPFN/actions/runs/18370843390/job/52333508672
Older versions of scikit-learn don't specify a maximum numpy version but
are incompatible with numpy>2. When using uv's lowest-direct mode, as in
the CI, scikit-learn is resolved to an old version (as it is a direct
dependency) but numpy is resolved to the newest specified compatible
version, which is >2. However, this is actually incompatible. Adding the
explicit numpy dependency is a trick which results in it being resolved
to <2. This doesn't help real users, but this issue has existed forever
and we haven't had any problems reported, and is common to all packages
depending on scikit-learn.

I selected the minimum version as the current minimum version required
by our dependencies, so this should have no effect on compatibility.