Fix potential race condition in InferenceEngineOnDemand.
If the number of workers used for the preprocessing is >1, then the
outputs could end up in a different order to the input configs, thus
each config might end up paired with the wrong output.

I don't think this bug could happen in practice because the n_workers
parameter is not connected, and we always use one worker. It's
reasonably likely we might want to use multiple workers in the near
future, so I don't remove n_workers.

I updated the tests to catch this race condition, and verified that the
test worked by disbaling the fix and setting the number of workers to >1.