Refactor to move restoring models into InferenceEngine.load_state().
This preparatory work for caching the models on device during inference.

At the same time, throw an exception if the model being saved uses a KV
cache, as this doesn't work (the KV cache was never saved, so I assume
you would get junk predictions after loading). It would be nicer and
fairly easy to make this work, but avoid it for now.