Don't run most of the MPS tests during PRs. (#634)
The MPS tests seem to be twice as slow as the CPU tests. Replace them
with a simple smoke test, and run most of them only on merge to main.

Add a "non_ci" marker to pytest, and mark the MPS tests with this.
Update the CI workflow to apply "not non_ci" during PRs.

Also, make sure tests that test on TABPFN_EXCLUDE_DEVICES override this
variable rather than assuming it's emtpy. Otherwise they break if the CI
sets it.

Part of RES-878