Test both v2 and v2.5 in consistency tests, and refactor the tests. (#639)
Include support for different reference predictions on different platforms/devices. For now, just include MacOS with apple silicon, as this is what's easy to generate on our laptops. I tried out a variety of different PyTorch and Python versions on my laptop, and they all seemed to result in the same reference predictions.

Also fix an in the CI workflow: it selected tests that matched "not non_ci", whereas the marker is actually "not_in_prs", so it should be "not not_in_prs". This resulted in more tests than necessary running during PRs.