Use github actions caching for the model files when in PRs. (#668)
Move the cache into the working directory so that it's consistent between OSes and accessible to hashFiles() (which is used to invalidate the cache).

The main aim is to allow PRs from forks to run the tests, as these don't have access to the HF_TOKEN and so can't download the model. This is isn't the ideal solution this problem, as the cache will be deleted [if no PRs are merged for 7 days](https://docs.github.com/en/actions/reference/workflows-and-actions/dependency-caching#usage-limits-and-eviction-policy), but it's a simple one. From [the docs](https://docs.github.com/en/actions/reference/workflows-and-actions/dependency-caching#restrictions-for-accessing-a-cache), I'm fairly confident that branches from forks will be able to access the cache populated in main, but I guess we will find out for sure...

Other advantages are avoiding spurious failures due to HF (which we do see sometimes), although only in PRs, and speeding up the PR CI by maybe 20 seconds.

The downside is that downloading the models will not be tested in PRs, but I don't restore the cache on main to test downloading there.