Try to fix model cache. (#670)
Relying on the tests to populate the cache didn't always seem to work on MacOS (sometimes the cache was empty). There was also a design flaw: we don't always use all the models when running in PRs, as some tests are skipped for speed. Thus, an entirely or partially empty cache sometimes got saved, which then broke PRs from forks.

Instead, explicitly use the "download_all_models()" function. Adapt it to skip models it already downloaded. Skip these steps entirely on main to continue to test to real download functionality.

This also doesn't fix the first-party compatibility test for tabpfn-extensions, as this seems to ignore the model cache dir, but I think that's fine for now (this check doesn't block submission).